[mm] feat: cleanup codes & bugs fixup
feat: moving varea pgmgr to anon.
Page manager is only a internal structure of anonymous mapping
currently. So it's moved to mm_anon.c now.
This patch also includes a bug fix of extra cleanup of varea in page
manager while doing varea_merge operations.

feat: remove restriction on page install

fixup: mm test case.
Using the assertion uassert_true(!(lwp_ref_dec(lwp) - 1)) instead of
uassert_true(!(lwp_ref_dec(lwp))) since the lwp_ref_dev returns the
previous value of the lwp reference
