+++ libcpu/aarch64/SConscript
@@ -12,9 +12,9 @@ bsp_path = Dir('#').abspath
 
 if not os.path.exists(bsp_path + "/link.lds"):
     Env['LINKFLAGS'] = Env['LINKFLAGS'].replace('link.lds', cwd + "/link.lds")
-    # fix the linker with crtx.o
     Preprocessing("link.lds.S", ".lds", CPPPATH=[bsp_path])
 
+# fix the linker with crtx.o
 Env['LINKFLAGS'] += ' -nostartfiles'
 
 # add common code files
